+++ src/main/java/br/com/zalf/prolog/webservice/commons/util/Now.java
@@ -1,6 +1,8 @@
 package br.com.zalf.prolog.webservice.commons.util;
 
 import java.sql.Timestamp;
+import java.time.Clock;
+import java.time.LocalDate;
 
 /**
  * Created on 04/03/2018
@@ -19,5 +21,9 @@ public final class Now {
 
     public static Timestamp timestampUtc() {
         return new Timestamp(utcMillis());
+    }
+
+    public static LocalDate localDate() {
+        return LocalDate.now(Clock.systemUTC());
     }
 }