+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/servico/relatorio/ServicoRelatorioResource.java
@@ -32,4 +32,21 @@ public final class ServicoRelatorioResource {
             throws RuntimeException {
         return outputStream -> service.getEstratificacaoServicosFechadosCsv(outputStream, codUnidade, dataInicial, dataFinal);
     }
+
+    @GET
+    @Path("/estratificacao-servicos-abertos/{codUnidade}/report")
+    public Report getEstratificacaoServicosAbertosReport(@PathParam("codUnidade") @Required final Long codUnidade,
+                                                         @QueryParam("dataInicial") @Required final String dataInicial,
+                                                         @QueryParam("dataFinal") @Required final String dataFinal) {
+        return service.getEstratificacaoServicosAbertosReport(codUnidade, dataInicial, dataFinal);
+    }
+
+    @GET
+    @Path("/estratificacao-servicos-abertos/{codUnidade}/csv")
+    public StreamingOutput getEstratificacaoServicosAbertosCsv(@PathParam("codUnidade") @Required final Long codUnidade,
+                                                               @QueryParam("dataInicial") @Required final String dataInicial,
+                                                               @QueryParam("dataFinal") @Required final String dataFinal)
+            throws RuntimeException {
+        return outputStream -> service.getEstratificacaoServicosAbertosCsv(outputStream, codUnidade, dataInicial, dataFinal);
+    }
 }